Add files via upload
This project is a console-based game where the compiler picks the random number by the use of rand() and srand() functions and the player guesses the number. It also displays the message based on whether the number is too high or too low. according to the number given by the user.